requirements.txt
faiss-cpu
sentence-transformers
numpy
pandas
openai
transformers
streamlit